# Optimize page rank contention