--- a/src/function/gds/gds.cpp
+++ b/src/function/gds/gds.cpp
@@ -31,14 +31,8 @@ std::shared_ptr<Expression> GDSAlgorithm::bindNodeOutput(Binder* binder,
 }
 
 std::shared_ptr<PathLengths> GDSAlgorithm::getPathLengthsFrontier(
-    processor::ExecutionContext* context, uint16_t val) {
-    auto tx = context->clientContext->getTransaction();
-    auto mm = context->clientContext->getMemoryManager();
-    auto graph = sharedState->graph.get();
-    auto pathLengths = std::make_shared<PathLengths>(graph->getNumNodesMap(tx), mm);
-    auto vc = std::make_unique<PathLengthsInitVertexCompute>(*pathLengths, val);
-    GDSUtils::runVertexCompute(context, graph, *vc);
-    return pathLengths;
+    processor::ExecutionContext* context, uint16_t initialVal) {
+    return PathLengths::getFrontier(context, sharedState->graph.get(), initialVal);
 }
 
 } // namespace function