--- a/src/function/gds/gds_frontier.cpp
+++ b/src/function/gds/gds_frontier.cpp
@@ -1,6 +1,7 @@
 #include "function/gds/gds_frontier.h"
-#include "processor/execution_context.h"
+
 #include "function/gds/gds_utils.h"
+#include "processor/execution_context.h"
 
 using namespace kuzu::common;
 
@@ -108,7 +109,8 @@ void PathLengths::pinNextFrontierTableID(common::table_id_t tableID) {
     nextFrontier.store(getMaskData(tableID), std::memory_order_relaxed);
 }
 
-std::shared_ptr<PathLengths> PathLengths::getFrontier(processor::ExecutionContext* context, graph::Graph* graph, uint16_t initialValue) {
+std::shared_ptr<PathLengths> PathLengths::getFrontier(processor::ExecutionContext* context,
+    graph::Graph* graph, uint16_t initialValue) {
     auto tx = context->clientContext->getTransaction();
     auto mm = context->clientContext->getMemoryManager();
     auto pathLengths = std::make_shared<PathLengths>(graph->getNumNodesMap(tx), mm);