--- a/src/function/gds/k_core_decomposition.cpp
+++ b/src/function/gds/k_core_decomposition.cpp
@@ -10,6 +10,7 @@
 #include "graph/graph.h"
 #include "processor/execution_context.h"
 #include "processor/result/factorized_table.h"
+#include "degrees.h"
 
 using namespace kuzu::binder;
 using namespace kuzu::common;
@@ -19,45 +20,6 @@ using namespace kuzu::graph;
 
 namespace kuzu {
 namespace function {
-
-using degree_t = uint64_t;
-static constexpr degree_t INVALID_DEGREE = UINT64_MAX;
-
-class Degrees {
-public:
-    Degrees(const table_id_map_t<offset_t>& numNodesMap, MemoryManager* mm) {
-        init(numNodesMap, mm);
-    }
-
-    void pinTable(table_id_t tableID) { degreeValues = degreeValuesMap.getData(tableID); }
-
-    void addDegree(offset_t offset, uint64_t degree) {
-        degreeValues[offset].fetch_add(degree, std::memory_order_relaxed);
-    }
-
-    void decreaseDegreeByOne(offset_t offset) {
-        degreeValues[offset].fetch_sub(1, std::memory_order_relaxed);
-    }
-
-    degree_t getValue(offset_t offset) {
-        return degreeValues[offset].load(std::memory_order_relaxed);
-    }
-
-private:
-    void init(const table_id_map_t<offset_t>& numNodesMap, MemoryManager* mm) {
-        for (const auto& [tableID, numNodes] : numNodesMap) {
-            degreeValuesMap.allocate(tableID, numNodes, mm);
-            pinTable(tableID);
-            for (auto i = 0u; i < numNodes; ++i) {
-                degreeValues[i].store(0, std::memory_order_relaxed);
-            }
-        }
-    }
-
-private:
-    std::atomic<degree_t>* degreeValues = nullptr;
-    ObjectArraysMap<std::atomic<degree_t>> degreeValuesMap;
-};
 
 class CoreValues {
 public:
@@ -126,22 +88,6 @@ public:
 private:
     Degrees* degrees;
     CoreValues* coreValues;
-};
-
-struct DegreeEdgeCompute : public EdgeCompute {
-    Degrees* degrees;
-
-    explicit DegreeEdgeCompute(Degrees* degrees) : degrees{degrees} {}
-
-    std::vector<nodeID_t> edgeCompute(nodeID_t boundNodeID, graph::NbrScanState::Chunk& chunk,
-        bool) override {
-        degrees->addDegree(boundNodeID.offset, chunk.size());
-        return {};
-    }
-
-    std::unique_ptr<EdgeCompute> copy() override {
-        return std::make_unique<DegreeEdgeCompute>(degrees);
-    }
 };
 
 struct RemoveVertexEdgeCompute : public EdgeCompute {
@@ -231,6 +177,7 @@ public:
           coreValue{coreValue}, numActiveNodes{numActiveNodes} {}
 
     bool beginOnTable(table_id_t tableID) override {
+        frontierPair->pinNextFrontier(tableID);
         degrees->pinTable(tableID);
         coreValues->pinTable(tableID);
         return true;
@@ -295,24 +242,16 @@ public:
         auto graph = sharedState->graph.get();
         auto numNodesMap = graph->getNumNodesMap(clientContext->getTransaction());
         auto degrees = Degrees(numNodesMap, mm);
+        DegreesUtils::computeDegree(context, graph, &degrees, ExtendDirection::BOTH);
         auto coreValues = CoreValues(numNodesMap, mm);
         auto currentFrontier = getPathLengthsFrontier(context, PathLengths::UNVISITED);
-        auto nextFrontier = getPathLengthsFrontier(context, 0);
+        auto nextFrontier = getPathLengthsFrontier(context, PathLengths::UNVISITED);
         auto frontierPair = std::make_unique<KCoreFrontierPair>(currentFrontier, nextFrontier,
             &degrees, &coreValues);
-        // Initialize starting nodes (all nodes) in the next frontier.
-        // When beginNewIteration, next frontier will become current frontier
-        frontierPair->setActiveNodesForNextIter();
-        frontierPair->getNextSparseFrontier().disable();
-        // Compute Degree
-        auto degreeEdgeCompute = std::make_unique<DegreeEdgeCompute>(&degrees);
-        auto computeState = GDSComputeState(std::move(frontierPair), std::move(degreeEdgeCompute),
-            sharedState->getOutputNodeMaskMap());
-        GDSUtils::runFrontiersUntilConvergence(context, computeState, graph, ExtendDirection::BOTH,
-            1 /* maxIters */);
         // Compute Core values
         auto removeVertexEdgeCompute = std::make_unique<RemoveVertexEdgeCompute>(&degrees);
-        computeState.edgeCompute = std::move(removeVertexEdgeCompute);
+        auto computeState = GDSComputeState(std::move(frontierPair), std::move(removeVertexEdgeCompute),
+            sharedState->getOutputNodeMaskMap());
         auto coreValue = 0u;
         auto numNodes = graph->getNumNodes(clientContext->getTransaction());
         auto numNodesComputed = 0u;