--- a/src/include/function/gds/gds_frontier.h
+++ b/src/include/function/gds/gds_frontier.h
@@ -194,7 +194,8 @@ public:
     void pinCurFrontierTableID(common::table_id_t tableID);
     void pinNextFrontierTableID(common::table_id_t tableID);
 
-    static std::shared_ptr<PathLengths> getFrontier(processor::ExecutionContext* context, graph::Graph* graph,uint16_t initialValue);
+    static std::shared_ptr<PathLengths> getFrontier(processor::ExecutionContext* context,
+        graph::Graph* graph, uint16_t initialValue);
 
 private:
     uint16_t getCurIter() { return curIter.load(std::memory_order_relaxed); }