# Description

Improve page rank contention by using backward adjacency list when propagating ranks.

On the same LDBC100 person know person, performance becomes

1 thread - 5.2s
2 thread  - 2.6s
4 thread - 2.3s

The previous benchmark number was

5s, 2.7s, 4s

Fixes # (issue)

# Contributor agreement

- [ ] I have read and agree to the [Contributor Agreement](https://github.com/kuzudb/kuzu/blob/master/CLA.md).