@andyfengHKU 2025-01-13
> Looks great to me! The refactoring is nice.
> 
> I have two todos in mind:
> 
> * Can you add a comment to the PageRank implementation to explain the backward traversing? I don't think that is intuitive without any comments.
> * Can you also show performance numbers before the optimization? It's clearer how much improvement this brings into.

Updated the benchmark number. I'll add comments in my next PR.